requirements.txt
absl-py==2.1.0
asttokens
backcall
Bottleneck
brotlipy==0.7.0
certifi==2022.5.18.1
cffi
charset-normalizer
colorama
cryptography
cycler
debugpy
decorator
detecto 
entrypoints 
executing 
fonttools==4.25.0
grpcio==1.68.1
idna 
ipykernel 
ipython 
jedi 
joblib 
jupyter-client 
jupyter-core 
kiwisolver 
Markdown==3.7
MarkupSafe==3.0.2
matplotlib 
matplotlib-inline 
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio 
numexpr 
numpy 
opencv-python==4.10.0.84
packaging 
pandas 
parso 
pickleshare 
Pillow==9.0.1
prompt-toolkit 
protobuf==4.21.2
psutil==6.1.1
pure-eval 
pycparser 
Pygments 
pyOpenSSL 
pyparsing 
PySocks 
python-dateutil 
pytz==2021.3
pywin32==302
PyYAML==6.0.2
pyzmq 
requests 
scikit-learn 
scipy 
seaborn 
sip==4.19.13
six 
skorch==1.0.0
soundfile==0.12.1
stack-data 
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl 
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado 
tqdm 
traitlets 
typing_extensions 
urllib3 
wcwidth 
Werkzeug==3.1.3
win-inet-pton 
wincertstore==0.2
